Assorted Fixes, Improvements for Hunter Support
Classes:  Block a 'pseudo-start' of combat for actions that are "in-air"
and projected to land.
+ Add Netherwinds to the list of Bloodlust buffs.

Core:  Initial implementation of "flicker protection."
+ Basically, if it sees that the addon's last several recommendations
have flickered between the newest actions and something else, it just
won't bother updating.  Will be iterating on this.

Events:  Spells that have been cast and are projected to land will
default to a 0.05 travel time.

State:  'query_time' will now incorporate latency into its
recommendation.  This will result in the addon making recommendations
based on a slightly later game-state.
+ Holy Power is treated like a regenerating resource when Liadrin's Fury
Unleashed is equipped and AW/Crusade is up.  THIS NEEDS TESTING.
+ Tried to improve ability "timeToReady" calculations for Survival
Hunters when using Spitting Cobra.