I should split things off more often.
General
-------
+ Trying out a new time-slice based engine instead of the
projection-based system.  The projection-based system has required too
many special accommodations (time scripts, hidden retesting, etc.).  The
major difference is that the addon will try to see what to recommend
now, and if it doesn't find something, it will just step forward a bit
in time and try again.  This is more like SimulationCraft, which tests
for an action and waits 0.1 seconds if no action is available.  The
addon takes a first 0.1s step and if it doesn't find anything there, it
starts looking at every quarter-GCD for a recommendation until it finds
one or runs out of time trying.  Work in progress.

+ The remnants of a new resource modeling system has been added, but
will likely be removed again soon.  It was build for the projection
system which is no longer used.

Classes
-------
+ Qualify some auras as "incapacitates" so that the addon will assume
you can't use abilities during that time frame.  (And by some, I mean
only Draught of Souls for now.)

Paladin
-------
+ Guarantee that Liadrin's grants you the first Holy Power from
AW/Crusade.

Warlock
-------
+ Rebuilt Soul Shard model; they don't regenerate in any way that we are
concerned with.

Scripts
-------
+ Legend has it that C++ "!0>1" and Lua "not 0 > 1" are not equivalent.
Fix up auto-conversion to make the addon use "not ( 0 > 1 )" where
applicable.

State
-----
+ Lots of resource modeling nonsense still in here.
+ Support 'in_flight' check (if an ability has been cast before the
addon simulation, but hasn't hit).