Elemental APL tweak.
Removed the BL up or time > 60 logic for Ascendance in the Elemental
APL.

Hekili:GetAbilityInfo() will return the ability ID, name, key, and
itemID (if applicable) for values returned by
Hekili_GetRecommendedAbility().

Hekili:GetDisplayByName() will return the display ID of a display based
on its name.

Hekili:GetActiveDisplays() returns a table of display indices based on
whether they are visible or not.  (Work in progress.)