Lower impact TTD calculations.
Paladin's judgment_override is back to latency x2.  May just cap it
instead.
Fix bug in Recheck function for Nightblade.
Try to tune Fury rage generation a bit more.
Purge the time-to-die DB when leaving combat.
Abilities that require a buff is up (or down) are unusable, not unknown
if the buff is down (or up).
Borrow emptyrivers' iterator for going through your whole party/raid
unit list.  Not actually using it yet, though.