Melee Range is 2 yards
Melee Range is actually `2` yards, not the `5` from in game tooltips (or the `7` that was in Hekili's code).

This is correct for all melee classes except Druids with Balance Affinity and Outlaw Rogues with Acrobatic Strikes.  For these, it doesn't seem to be as simple as adding 3 yards as `>= 5` yards is too short and `> 5` is too long so an ability will probably have to be checked in these circumstances.