Improving Bartender 4 Support
Added a rather simple fix for bar paging when going into stealth as a rogue. This can be adapted for use with druids as well.

This (just like the previous implementation) doesn't handle the customizable Bartender 4 bar paging at all. The new code simply maps spells in bar 7 (default action bar page for stealth) onto the key binds for bar 1 when the player is a rogue.

With this fix Hekili displays proper keybindings when in stealth. This breaks bar 7 for normal (non-stealth) use when playing rogue.