conduit.nimble_fingers.mod fix
The logic behind "conduit.nimble_fingers" is to decrease the value (https://github.com/Hekili/hekili/blob/shadowlands/Shadowlands/SoulbindConduits.lua Line 247) so we get less energy spent. But here if we substract the already negative value we actually increase the energy cost of the spell. Here is a fix.