Remove mad bombardier buff on cast
With sufficiently high haste (~30%) then even in melee range you will be able to press another global before the mad bombardier buff is consumed (on impact). Remove the buff on cast so suggestions cant be made based on it while the projectile is in the air.
Still only gain wildfire bomb charges on impact.