[APL Tweak] Outlaw
As Simcraft did not make any adjustments into prepatch version of the APL and only improved DF version of it. I want to suggest implementing some changes to represent both prepatch and future DF changes of the APL considering legendaries, talents, etc.
Link to the simc APL https://github.com/simulationcraft/simc/blob/dragonflight/engine/class_modules/apl/rogue/outlaw_df.simc


Adjusted RTB reroll logic to actually not reroll Broadside and True bearing during Loaded Dice buff (line 21 of the DF APL)

Finish condition typo fix

Echoing Reprimand should not be used during dreadblades window. (Line 44 of the DF APL)

Ambush adjustment (Line 45 and line 53 of the DF APL)

Pool resources for Dreadblades as it costs 50 energy which is more than Sinister Strike and ofter i came up in a situation that i have no CD on dreadblade (and it should be cast on CD) but not enough energy to cast it and addon suggested me to just cast a filler (Sinister Strike) instead of waiting for an energy tick.

Vansich adjustments with Invigorating Shadowdust legendary and not a Venthyr covenant. This is needed or else in the non-SL related content no conditions for vanish are met and it will not be used at all