Pal Ret: Add a new option that changes the way Seraphim is played.
This allows for a more intense damage window every minute. Of course, this slightly changes the rotation predicted by simulation, and results in slightly lower overall damage in long fights, which is why it's an optional, toggleable setting. I wanted to propose it in case it might be suitable for some people.