Update EvokerPreservation.lua
Readded runHandler Hook for actions, incorrectly left out of prior merge of Devastation work, which caused major issues with multiple functions, especially while empowering a spell.

Commented out Cycle of Life handling in function for Emerald Blossom until fully ready to implement.  This was causing the APL to hang when selected and when EB was used.