Added Consecration Buff for Strength of Conviction
Consecration: Strength of Conviction talent gives a buff (188370) which you can lose by walking out of it, so buff.consecration.up won't detect it using the other code handling that.  I added this buff to detect only the you're-standing-in-it-now part as that might want to trigger an early recast for folks.  I confirmed that the spell ID is the same whether you have 1 or 2 points in the talent.