feat: extract a caption for an action from its comment
In the APL profile, when pairing up a comment as the description for the
action in the subsequent line, if the comment has the format

  <Caption> :: <Description>

then `<Caption>` is used as the caption for the action, and
`<Description>` is used as the description for the action. Otherwise the
comment is used unmodified as the description for the action.