Update MageArcane.lua
A few more fixes related to the previous one, since Magis Spark is not a self buff. 

Removed also the id on magis_spark since what is applied is Touch of the Magi as a debuff. Unless that needs to be kept. The spell ID referenced originally, https://www.wowhead.com/spell=450004/magis-spark, isn't actually a buff applied to the caster at all. It's the debuff Touch of the Magi that is modified to behave differently.

All that said, none of these changes still seem to work to make it so that Arcane Blast is recommended ONCE touch of the magi is applied when talented into Magis Spark, while in AOE.

This snapshot shows the sequence recommended, but when Touch of the Magi is actually pressed, it doesn't seem like the debuff.magis_spark.arcane_blast or even other ones, is applied at all in reality.