Fix #3641
Basically, the issue is that Magis Spark is not actually a buff applied to the player. The code originally references this https://www.wowhead.com/spell=450004/magis-spark.

The talent simply modifies the Touch of the Magi debuff. So the changes here take that into account. Applying the three ability debuffs (magis_spark_arcane_barrage, magis_spark_arcane_missles, and magis_spark_arcane_blast) now apply properly when Touch of the Magi is cast, and Arcane Blast is recommended once even while in AOE range.