Shadow Crash Opener
No reason to have shadowcrash be reliant on a boss condition in precombat opener. In the SIMC apl it is simply "not dungeon slice". Having it show up in the AoE opener is natural and abides by the guides, as well as AoE sims.

Fix for: https://github.com/Hekili/hekili/issues/3644