refactor: push some addon settings into the priorities instead
Change some Protection Warrior settings so they are optionally checked
within the priorities instead of directly in the abilities' `usable`
functions. In particular, don't force the addon to enforce the settings
for the following spells so the usage conditions can be set directly in
the priorities:

- Last Stand
- Rallying Cry
- Shield Wall

Some of these settings made sense in previous expansions but
current SimulationCraft profiles already try to take those extra
conditions into account.

Add new expressions that can be used in action modifiers:

- `last_stand_damage_taken`
- `last_stand_health_pct`
- `rallying_cry_damage_taken`
- `rallying_cry_health_pct`
- `shield_wall_damage_taken`
- `shield_wall_health_pct`

Remove the settings that checked for "damage taken *and* health
conditionals" as they can be individually checked within the priorities
through the exposed expressions.

No changes are made in the default protection warrior priority as Last
Stand and Shield Wall are both used offensively in the default priority.