feat: only fade if the recommended ability can't be spam-cast
When the recommended ability can't be spam-cast because it requires
waiting past the end of the GCD or the current spellcast, then fade the
ability as unusable if the appropriate setting is toggled.

Remove a commented code block that is a duplicate of code in the "delay"
check as we directly re-use the value that is stored in
`b.EarliestTime`.