fix: manage Marked for Execution debuff from Slayer's Strike
Modify Arms and Fury warrior modules to manage the Marked for Execution
stacking debuff from Slayer's Strike.

Marked For Execution is applied by Slayer's Strike and stacks a maximum of
three times and is removed only when Execute is cast.

With the Imminent Demise talent, consuming Sudden Death applies Imminent
Demise and stacks a maximum of three times and is removed by Bladestorm.

Use the correct name for the debuff (also used in SimC) as
`marked_for_execution` and allow it to be tracked via `active_dot`.