Spriest - Entropic Rift
Improve Entropic Rift calculations. Is is inherently linked to voidheart which is a real, trackable buff with an ID. So alias it with entropic rift and use voidheart for all calculations. Removes need for reset_precast calculating as well.

void torrent refreshes the 8 second buff on every tick, and void blast extends it by a second each cast, up to a max of 3 seconds per window.