Fix unknown buff "aethervision" in Arcane Mage module
Uncommented the aethervision buff definition and related handler code in MageArcane.lua to resolve the error "Unknown buff in [Arcane:sunfury_aoe:6]: aethervision" that was occurring when the addon tried to reference this buff in the priority list conditions. This ensures proper functionality of the Arcane Mage module when using the sunfury_aoe action list.