Re-organise Outlaw, add Coup de Grace logic
Added Coup de Grace logic, as it will be needed as a separate spell for Trickster. (During my internal testing the addon treated Coup de Grace exclusively as a Dispatch, which in some APL cases is not correct.)
There is no Coup de Grace buff, we have to manually cound stacks of Escalating Blade. The delay in consuming of the debuff is needed to use a "double Coup de Grace" technique, which is technically a bug, but provides around 5% dps increase.

Added Disorienting Strikes buff, so it can be consumed with Sinisteer Strike and Ambush.
--TO DO: fix Sinister Strike logic of applying Unseen Blade, may be internal CD of 20s? Also need to change Ambush handler in Assassination.lua to consume Disorienting Strikes.

Re-organised the file to follow the previous logic, but just with fewer repeated table lookups.
Upvalued GetTime as getTime, FindPlayerAuraByID, etc.
Minimized repeated _G or math.XXX references.