Register Action Packs without Export Strings ### Goal Do not require action packs to be exported from in-game to be registered as a default.
This also eliminates the need for the Priorities folder and ensures there is one source of truth for a registered default priority.

### Implementation
This implementation lets you provide the APL as a long text string followed by a table of additional arguments.

`spec:RegisterPriority( name, version, apl, args )`

The additional args are in table format to avoid having to nil additional pack elements you don't want to provide.

Valid `args`:
* author (string)
* desc (string)
* source (string)
* hidden (bool)

### TODO
- [ ] Don't `ImportSimcAPL` unless we know we need to refresh due to version discrepancy or manual refresh (reduces load time)
- [ ] Revise source to link to priority source
- [ ] Add commit to args
- [ ] Use source + commit to show link to commit in Options
- [ ] ???